docker_app/requirements.txt
torch
torchvision
streamlit
numpy
diffusers
pillow
transformers
requests
